+++ app/src/main/java/br/com/laboratorio/hemope/Aliquota/AliquotaFragment.java
@@ -10,6 +10,7 @@ import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v4.view.ViewPager;
@@ -47,17 +48,18 @@ import br.com.laboratorio.hemope.Model.LocalProcedencia;
 import br.com.laboratorio.hemope.Model.Paciente;
 import br.com.laboratorio.hemope.Model.TipoAmostra;
 import br.com.laboratorio.hemope.R;
+import br.com.laboratorio.hemope.Util;
 import br.com.laboratorio.hemope.View.SlidingTabLayout;
 
 
 public class AliquotaFragment extends Fragment {
 
-        View aliquotaView;
+        static View aliquotaView;
         ViewPager viewPager;
         SlidingTabLayout mSlidingTabLayout;
-        DownloadAliquotaTask task;
 
-        Itens itens;
+
+        static Itens _itens;
         ProgressDialog progressDialog;
 
         public AliquotaFragment() {
@@ -70,7 +72,7 @@ public class AliquotaFragment extends Fragment {
                 intent.putExtra("SCAN_MODE", "QR_CODE_MODE");
                 startActivityForResult(intent, 0);
             } catch (ActivityNotFoundException anfe) {
-                showDialog(getActivity(), "Sem Scanner Encontrado!", "Baixar um Scanner agora?", "Sim", "Não").show();
+                Util.showDialog(getActivity(), "Sem Scanner Encontrado!", "Baixar um Scanner agora?", "Sim", "Não").show();
             }
         }
 
@@ -89,31 +91,6 @@ public class AliquotaFragment extends Fragment {
 
             ((AcaoPrincipalActivity) activity).onSectionAttached(
                     getArguments().getInt(ARG_SECTION_NUMBER));
-        }
-
-
-
-        private static AlertDialog showDialog(final Activity act, CharSequence title, CharSequence message, CharSequence buttonYes, CharSequence buttonNo) {
-            AlertDialog.Builder downloadDialog = new AlertDialog.Builder(act);
-            downloadDialog.setTitle(title);
-            downloadDialog.setMessage(message);
-            downloadDialog.setPositiveButton(buttonYes, new DialogInterface.OnClickListener() {
-                public void onClick(DialogInterface dialogInterface, int i) {
-                    Uri uri = Uri.parse("market://search?q=pname:" + "com.google.zxing.client.android");
-                    Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-                    try {
-                        act.startActivity(intent);
-                    } catch (ActivityNotFoundException anfe) {
-
-                    }
-                }
-            });
-            downloadDialog.setNegativeButton(buttonNo, new DialogInterface.OnClickListener() {
-                public void onClick(DialogInterface dialogInterface, int i) {
-
-                }
-            });
-            return downloadDialog.show();
         }
 
 
@@ -144,12 +121,12 @@ public class AliquotaFragment extends Fragment {
                 case R.id.novaAlocacaoAliquota:
 
                     try {
-                        if(itens.aliquota.idAliquota != null) {
+                        if(_itens.aliquota.idAliquota != null) {
                             FragmentTransaction transaction = getFragmentManager()
                                     .beginTransaction();
                             Bundle argsaLocacaoFragment = new Bundle();
                             argsaLocacaoFragment.putInt(ARG_SECTION_NUMBER, 5);
-                            argsaLocacaoFragment.putSerializable("itens", itens);
+                            argsaLocacaoFragment.putSerializable("itens", _itens);
                             Fragment novaAlocacaoFragment = new AlocacaoFragment();
 
                             novaAlocacaoFragment.setArguments(argsaLocacaoFragment);
@@ -184,7 +161,10 @@ public class AliquotaFragment extends Fragment {
                          try{
                              //Verifico se é um número
                              if(Integer.parseInt(idAliquota) > 0) {
-                                    consultarAliquota(idAliquota);
+
+                                 Util.DownloadTask downloadTask = new Util.DownloadTask("Carregando","Aguarde dados da Aliquota...","aliquota",_itens,getActivity());
+                                 downloadTask.execute("https://www.dropbox.com/s/10enzhvm6lnb0j6/aliquota.json?dl=1");
+
                              }else{
                                  Toast.makeText(getActivity(), "ID da Aliquota Inválido", Toast.LENGTH_LONG).show();
                              }
@@ -200,53 +180,10 @@ public class AliquotaFragment extends Fragment {
             }
         }
 
-        public void consultarAliquota(String idAliquota){
 
 
-            task = new DownloadAliquotaTask();
-            task.execute(idAliquota);
-        }
 
-
-        class DownloadAliquotaTask extends AsyncTask<String, Void, Itens> {
-
-            @Override
-            protected Itens doInBackground(String... pesquisa) {
-                OkHttpClient client = new OkHttpClient();
-
-                Request request = new Request.Builder()
-                        .url("https://www.dropbox.com/s/10enzhvm6lnb0j6/aliquota.json?dl=1")
-                        .build();
-
-                try {
-                    Response response = client.newCall(request).execute();
-                    String json = response.body().string();
-
-                    Gson gson = new Gson();
-                    itens = gson.fromJson(json, Itens.class);
-
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                return itens;
-            }
-
-            @Override
-            protected void onPreExecute(){
-                progressDialog = ProgressDialog.show(getActivity(), "Aguarde...", "Carregando dados da Aliquota...", true);
-                progressDialog.setCancelable(false);
-            }
-
-            @Override
-            protected void onPostExecute(Itens aliquota) {
-                super.onPostExecute(aliquota);
-                progressDialog.dismiss();
-                preencherActivity();
-            }
-    }
-
-
-    public void preencherActivity(){
+    public static void preencherActivityAliquota(Itens itens, FragmentActivity context){
 
         Paciente paciente = new Paciente();
         Aliquota aliquota = new Aliquota();
@@ -261,6 +198,7 @@ public class AliquotaFragment extends Fragment {
         LocalProcedencia localProcedencia = new LocalProcedencia();
 
     try {
+        _itens = itens;
         aliquota = itens.aliquota;
         amostra = itens.aliquota.amostra;
         alocacao = itens.aliquota.alocacao;
@@ -318,20 +256,12 @@ public class AliquotaFragment extends Fragment {
         txtSiglaDiagnostico.setText("Sigla Diagnóstico: " + diagnostico.sigla);
         txtCodCid.setText("Cod. CID: " + cid.codigo);
         txtDescricaoCid.setText("Sigla CID: " + cid.descricao);
-        Toast.makeText(getActivity(), "Carregamento Concluído.", Toast.LENGTH_SHORT).show();
+        Toast.makeText(context, "Carregamento Concluído.", Toast.LENGTH_SHORT).show();
 
     }catch (Exception e){
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-        builder.setTitle("Conexão")
-                .setMessage("Erro ao tentar se conectar com os Servidores.")
-                .setCancelable(false)
-                .setNegativeButton("OK", new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        dialog.cancel();
-                    }
-                });
-        AlertDialog alert = builder.create();
-        alert.show();
+
+        Util.exibirMensagem("conexão","Erro ao tentar se conectar com os Servidores.",context);
+
     }
 
     }
